Rimossi usi di JSONObject (manca da sostiture mymarker e myjson)

+++ app/src/main/java/com/example/maptry/model/liveevents/LiveEvent.kt
@@ -7,5 +7,5 @@ data class LiveEvent(
     val longitude: Double,
     val name: String,
     val owner: String,
-    val expirationDate: Int
+    val expirationDate: Long
 )
